refactor GoalScreen
